Separate Sensor reading from reflow code in effort to reduce variable spagetti in reflow code
